refactor: Организовал страницы и стили в папки
