+++ Syncro.Client/syncro-frontend/src/Styles/Register.css
